+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Features:
 * Support MATE Terminal version & terminal font detection
 * Add `--no-buffer` option for easier debugging. CMake option `ENABLE_BUFFER` is removed and always enabled.
 * Support `--*-key-color` option to change the key color of specified module
+* Support `--colors-symbol`
 
 # 1.11.3